anomaly_detectorr/requirements.txt
scapy
psutil
PyYAML
rich
requests
